chore: moved FixReferences() from MockExtension to MockManager
