+++ JotunnLib/Managers/MockSystem/MockManager.cs
@@ -1,4 +1,7 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Reflection;
+using Jotunn.Utils;
 using UnityEngine;
 using Object = UnityEngine.Object;
 
@@ -10,16 +13,17 @@ namespace Jotunn.Managers
     internal class MockManager : IManager
     {
         private static MockManager _instance;
+
         /// <summary>
         ///     The singleton instance of this manager.
         /// </summary>
         public static MockManager Instance => _instance ??= new MockManager();
-        
+
         /// <summary>
         ///     Hide .ctor
         /// </summary>
         private MockManager() {}
-        
+
         /// <summary>
         ///     Legacy ValheimLib prefix used by the Mock System to recognize Mock gameObject that must be replaced at some point.
         /// </summary>
@@ -30,7 +34,7 @@ namespace Jotunn.Managers
         ///     Prefix used by the Mock System to recognize Mock gameObject that must be replaced at some point.
         /// </summary>
         public const string JVLMockPrefix = "JVLmock_";
-        
+
         /// <summary>
         ///     Internal container for mocked prefabs
         /// </summary>
@@ -85,13 +89,14 @@ namespace Jotunn.Managers
                 Logger.LogWarning($"Could not create mock for prefab {prefabName} of type {typeof(T)}");
                 return null;
             }
+
             mock.name = name;
 
             Logger.LogDebug($"Mock {name} created");
 
             return mock;
         }
-        
+
         /// <summary>
         ///     Will try to find the real vanilla prefab from the given mock
         /// </summary>
@@ -132,7 +137,7 @@ namespace Jotunn.Managers
                                 unityObjectName.Substring(0, unityObjectName.Length - materialInstance.Length);
                         }
                     }
-                    
+
                     Object ret = PrefabManager.Cache.GetPrefab(mockObjectType, unityObjectName);
 
                     if (!ret)
@@ -147,5 +152,138 @@ namespace Jotunn.Managers
             return null;
         }
 #pragma warning restore CS0618
+
+        // Thanks for not using the Resources folder IronGate
+        // There is probably some oddities in there
+        internal static void FixReferences(object objectToFix, int depth)
+        {
+            // This is totally arbitrary.
+            // I had to add a depth because of call stack exploding otherwise
+            if (depth == 3 || objectToFix == null)
+            {
+                return;
+            }
+
+            depth++;
+
+            var type = objectToFix.GetType();
+            ClassMember classMember = ClassMember.GetClassMember(type);
+
+            foreach (var member in classMember.members)
+            {
+                // Special treatment for DropTable, its a List of struct DropData
+                // Maybe there comes a time when I am willing to do some generic stuff
+                // But mono did not implement FieldInfo.GetValueDirect()
+                if (member.MemberType == typeof(DropTable))
+                {
+                    var drops = ((DropTable)member.GetValue(objectToFix)).m_drops;
+
+                    for (int i = 0; i < drops.Count; i++)
+                    {
+                        var drop = drops[i];
+                        var realPrefab = MockManager.GetRealPrefabFromMock(drop.m_item, typeof(GameObject));
+                        if (realPrefab)
+                        {
+                            drop.m_item = (GameObject)realPrefab;
+                        }
+
+                        drops[i] = drop;
+                    }
+
+                    continue;
+                }
+
+                if (member.IsUnityObject && member.HasGetMethod)
+                {
+                    var target = (Object)member.GetValue(objectToFix);
+                    var realPrefab = MockManager.GetRealPrefabFromMock(target, member.MemberType);
+                    if (realPrefab)
+                    {
+                        member.SetValue(objectToFix, realPrefab);
+                    }
+                }
+                else
+                {
+                    if (member.IsEnumeratedClass && member.IsEnumerableOfUnityObjects)
+                    {
+                        var isArray = member.MemberType.IsArray;
+                        var isList = member.MemberType.IsGenericType && member.MemberType.GetGenericTypeDefinition() == typeof(List<>);
+                        var isHashSet = member.MemberType.IsGenericType && member.MemberType.GetGenericTypeDefinition() == typeof(HashSet<>);
+
+                        if (!(isArray || isList || isHashSet))
+                        {
+                            Logger.LogWarning($"Not fixing potential mock references for field {member.MemberType.Name} : {member.MemberType} is not supported.");
+                            continue;
+                        }
+
+                        var currentValues = (IEnumerable<Object>)member.GetValue(objectToFix);
+                        if (currentValues != null)
+                        {
+                            var list = new List<Object>();
+                            foreach (var unityObject in currentValues)
+                            {
+                                var realPrefab = MockManager.GetRealPrefabFromMock(unityObject, member.EnumeratedType);
+                                list.Add(realPrefab ? realPrefab : unityObject);
+                            }
+
+                            if (list.Count > 0)
+                            {
+                                MethodInfo cast = ReflectionHelper.Cache.EnumerableCast;
+                                MethodInfo castT = cast.MakeGenericMethod(member.EnumeratedType);
+                                object correctTypeList = castT.Invoke(null, new object[] { list });
+
+                                if (isArray)
+                                {
+                                    var toArray = ReflectionHelper.Cache.EnumerableToArray;
+                                    var toArrayT = toArray.MakeGenericMethod(member.EnumeratedType);
+
+                                    var array = toArrayT.Invoke(null, new object[] { correctTypeList });
+                                    member.SetValue(objectToFix, array);
+                                }
+                                else if (isList)
+                                {
+                                    var toList = ReflectionHelper.Cache.EnumerableToList;
+                                    var toListT = toList.MakeGenericMethod(member.EnumeratedType);
+
+                                    var newList = toListT.Invoke(null, new object[] { correctTypeList });
+                                    member.SetValue(objectToFix, newList);
+                                }
+                                else if (isHashSet)
+                                {
+                                    var hash = typeof(HashSet<>).MakeGenericType(member.EnumeratedType);
+
+                                    var newHash = Activator.CreateInstance(hash, correctTypeList);
+                                    member.SetValue(objectToFix, newHash);
+                                }
+                            }
+                        }
+                    }
+                    else if (member.IsEnumeratedClass)
+                    {
+                        var isDict = member.MemberType.IsGenericType && member.MemberType.GetGenericTypeDefinition() == typeof(Dictionary<,>);
+                        if (isDict)
+                        {
+                            Logger.LogWarning($"Not fixing potential mock references for field {member.MemberType.Name} : Dictionary is not supported.");
+                            continue;
+                        }
+
+                        var currentValues = (IEnumerable<object>)member.GetValue(objectToFix);
+                        if (currentValues == null)
+                        {
+                            continue;
+                        }
+
+                        foreach (var value in currentValues)
+                        {
+                            FixReferences(value, depth);
+                        }
+                    }
+                    else if (member.IsClass && member.HasGetMethod)
+                    {
+                        FixReferences(member.GetValue(objectToFix), depth);
+                    }
+                }
+            }
+        }
     }
 }
